scripts/requirements.txt
paho-mqtt==1.6.1
influxdb==5.3.1
tabulate==0.9.0
